backend/requirements.txt
# FastAPI 核心
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 資料庫
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 認證
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# AI/ML 服務
mlx-whisper>=0.4.0
requests>=2.31.0

# 說話者辨識
pyannote.audio>=3.1.0
torch>=2.0.0

# 匯出功能
reportlab>=4.0.0
python-docx>=1.1.0

# 工具
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0
greenlet>=3.0.0
